fix `AttributeError: 'collections.OrderedDict' object has no attribute 'float'` problem

+++ detect.py
@@ -58,7 +58,8 @@ def detect(save_image=False):
     model = YOLO(args.config_file).to(device)
 
     # Load weight
-    model.load_state_dict(torch.load(weights, map_location=device)["state_dict"]).float()
+    model.load_state_dict(torch.load(weights, map_location=device)["state_dict"])
+    model.float()
     model.eval()
 
     # Second-stage classifier
